Move voice input button
