+++ components/Chat/ChatInput.tsx
@@ -26,6 +26,7 @@ import HomeContext from '@/pages/api/home/home.context';
 import TokenCounter from './components/TokenCounter';
 
 import EnhancedMenu from '../EnhancedMenu/EnhancedMenu';
+import VoiceInputButton from '../VoiceInput/VoiceInputButton';
 import { PromptList } from './PromptList';
 import { VariableModal } from './VariableModal';
 
@@ -336,12 +337,6 @@ export const ChatInput = ({
             }
           `}
         >
-          <button
-            className="absolute left-2 top-2 rounded-sm p-1 text-neutral-800 opacity-60 dark:bg-opacity-50 dark:text-neutral-100 cursor-default"
-            onKeyDown={(e) => {}}
-          >
-            {getPluginIcon(currentMessage?.pluginId)}
-          </button>
 
           <EnhancedMenu
             ref={menuRef}
@@ -349,32 +344,44 @@ export const ChatInput = ({
             setIsFocused={setIsFocused}
           />
 
-          <textarea
-            ref={textareaRef}
-            className={`
-              m-0 w-full transition-all resize-none border-0 bg-transparent pt-3 pr-8 pl-10 text-black dark:bg-transparent dark:text-white outline-none
-              ${ isSpeechRecognitionActive ? 'z-[1100] pointer-events-none' : '' }
-            `}
-            style={{
-              marginBottom: `${
-                isCloseToTokenLimit || isOverTokenLimit ? '2.2' : '0.75'
-              }rem `,
-              resize: 'none',
-              bottom: `${textareaRef?.current?.scrollHeight}px`,
-              maxHeight: '400px',
-              overflow: `${
-                textareaRef.current && textareaRef.current.scrollHeight > 400
-                  ? 'auto'
-                  : 'hidden'
-              }`,
-            }}
-            placeholder={t('Type a message ...') || ''}
-            value={content}
-            rows={1}
-            onKeyUp={(e) => setIsTyping(e.nativeEvent.isComposing)}
-            onChange={handleChange}
-            onKeyDown={handleKeyDown}
-          />
+          <div className="flex items-start">
+            <div className="flex items-center pt-1 pl-1">
+              <VoiceInputButton />
+              <button
+                className="rounded-sm p-1 text-neutral-800 opacity-60 dark:bg-opacity-50 dark:text-neutral-100 cursor-default"
+                onKeyDown={(e) => {}}
+              >
+                {getPluginIcon(currentMessage?.pluginId)}
+              </button>
+            </div>
+
+            <textarea
+              ref={textareaRef}
+              className={`
+                m-0 w-full transition-all resize-none border-0 bg-transparent pt-3 pr-8 pl-2 text-black dark:bg-transparent dark:text-white outline-none
+                ${ isSpeechRecognitionActive ? 'z-[1100] pointer-events-none' : '' }
+              `}
+              style={{
+                marginBottom: `${
+                  isCloseToTokenLimit || isOverTokenLimit ? '2.2' : '0.75'
+                }rem `,
+                resize: 'none',
+                bottom: `${textareaRef?.current?.scrollHeight}px`,
+                maxHeight: '400px',
+                overflow: `${
+                  textareaRef.current && textareaRef.current.scrollHeight > 400
+                    ? 'auto'
+                    : 'hidden'
+                }`,
+              }}
+              placeholder={t('Type a message ...') || ''}
+              value={content}
+              rows={1}
+              onKeyUp={(e) => setIsTyping(e.nativeEvent.isComposing)}
+              onChange={handleChange}
+              onKeyDown={handleKeyDown}
+            />
+          </div>
 
           <TokenCounter
             className={` ${
