+++ components/VoiceInput/VoiceInputActiveOverlay.tsx
