+++ pages/api/home/home.tsx
@@ -51,7 +51,7 @@ import { AuthModel } from '@/components/User/AuthModel';
 import { ProfileUpgradeModel } from '@/components/User/ProfileUpgradeModel';
 import { SurveyModel } from '@/components/User/SurveyModel';
 import { UsageCreditModel } from '@/components/User/UsageCreditModel';
-import VoiceInputActiveOverlay from '@/components/EnhancedMenu/VoiceInputActiveOverlay';
+import VoiceInputActiveOverlay from '@/components/VoiceInput/VoiceInputActiveOverlay';
 
 import HomeContext from './home.context';
 import { HomeInitialState, initialState } from './home.state';
